Users/shiella/OneDrive/Desktop/MESIAS-PROJECT-main/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pyjwt==2.8.0
bcrypt==4.1.1
Pillow==10.1.0
